Refactoring

+++ src/evaluation/mobility.rs
@@ -8,10 +8,10 @@ pub fn evaluate(board: &Bitboard, white_attack_mask: &mut u64, black_attack_mask
 }
 
 fn evaluate_color(board: &Bitboard, color: u8, attack_mask: &mut u64) -> i16 {
-    let knight_mobility = get_piece_mobility::<KNIGHT>(board, color, attack_mask) as i16;
-    let bishop_mobility = get_piece_mobility::<BISHOP>(board, color, attack_mask) as i16;
-    let rook_mobility = get_piece_mobility::<ROOK>(board, color, attack_mask) as i16;
-    let queen_mobility = get_piece_mobility::<QUEEN>(board, color, attack_mask) as i16;
+    let knight_mobility = get_piece_mobility::<KNIGHT>(board, color, attack_mask);
+    let bishop_mobility = get_piece_mobility::<BISHOP>(board, color, attack_mask);
+    let rook_mobility = get_piece_mobility::<ROOK>(board, color, attack_mask);
+    let queen_mobility = get_piece_mobility::<QUEEN>(board, color, attack_mask);
 
     let game_phase = board.get_game_phase();
     let opening_score = (knight_mobility + bishop_mobility + rook_mobility + queen_mobility) * unsafe { MOBILITY_OPENING };
